chore(core): rename commentsSetup to addOnDataset/setup
