+++ packages/sanity/src/core/studio/addOnDataset/index.ts
@@ -1,0 +1,1 @@
+export * from './setup'
