+++ packages/sanity/src/structure/comments/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from '../../../../core/studio/commentsSetup/useCommentsSetup'
 export * from './use-comment-operations'
 export * from './useComments'
 export * from './useCommentsEnabled'
